[SDR-369] refactor: 스타일 관련 파일 style 폴더 내로 이동
- import order ESLint 설정 제거
