+++ fe/src/components/StoreDetail/StorePictures/StorePictures.tsx
@@ -1,5 +1,5 @@
 import { DEFAULT_IMG, STORE } from 'constants/dummyData';
-import { STORE_DETAIL } from 'constants/size';
+import { STORE_DETAIL } from 'styles/size';
 import { TEXT } from 'constants/text';
 import styled from 'styled-components';
 import { flexLayoutMixin } from 'utils/utils';
