+++ fe/src/pages/ReviewDetail/ReviewDetail.styled.ts
@@ -1,4 +1,4 @@
-import { STYLE } from 'constants/style';
+import { STYLE } from 'styles/style';
 import styled from 'styled-components';
 import { flexLayoutMixin } from 'utils/utils';
